Support for MS-MPI and Windows
This commit adds support for the MS-MPI implementation, primarily
targeting Windows. I add some code to discover for MS-MPI that only
run on Windows, and moved the other MPI probing to a separate probe
function for Unix.

I inject an msmpi cfg into both mpi-sys and rsmpi. This may be useful to
either surface msmpi specifically functionality or, like I do in this
commit, to disable certain features that are missing from MS-MPI.

I noticed two issues when building with the MS-MPI headers:
1. MPI_Comm_create_group was missing.
2. Bindgen does not seem to pick up the MPI_User_function definition

For both of these issues, I just removed the APIs that called these
functions when using MS-MPI. This may not be the best method.

I also disabled portions of any tests that rely on these
functionalities.

I recommend building without default features on Windows since libffi is
not available by default.

I also swapped the gcc crate with cc, as that seems to be
the stable version of gcc.
