+++ build-probe-mpi/Cargo.toml
@@ -12,4 +12,6 @@ categories = [ "development-tools::build-utils" ]
 license = "MIT/Apache-2.0"
 
 [dependencies]
+
+[target.'cfg(unix)'.dependencies]
 pkg-config = "0.3"
