+++ build-probe-mpi/src/lib.rs
@@ -8,7 +8,7 @@
 
 //! Probe an environment for an installed MPI library
 //!
-//! Probing is done in several steps:
+//! Probing is done in several steps on Unix:
 //!
 //! 1. Try to find an MPI compiler wrapper either from the environment variable `MPICC` or under
 //!    the name `mpicc` then run the compiler wrapper with the command line argument `-show` and
@@ -16,17 +16,20 @@
 //! 2. Query the `pkg-config` database for an installation of `mpich`.
 //! 3. Query the `pkg-config` database for an installation of `openmpi`.
 //!
+//! On Windows, only MS-MPI is looked for. The MSMPI_INC and MSMPI_LIB32/64 environment variables
+//! are expected.
+//!
 //! The result of the first successful step is returned. If no step is successful, a list of errors
 //! encountered while executing the steps is returned.
 
+#[cfg(unix)]
 extern crate pkg_config;
 
-use std::env;
-use std::error::Error;
-use std::path::PathBuf;
-use std::process::Command;
+mod os;
 
-use pkg_config::Config;
+pub use os::probe;
+
+use std::path::PathBuf;
 
 /// Result of a successfull probe
 #[derive(Clone, Debug)]
@@ -40,88 +43,4 @@ pub struct Library {
     /// The version of the MPI library
     pub version: String,
     _priv: (),
-}
-
-impl From<pkg_config::Library> for Library {
-    fn from(lib: pkg_config::Library) -> Self {
-        Library {
-            libs: lib.libs,
-            lib_paths: lib.link_paths,
-            include_paths: lib.include_paths,
-            version: lib.version,
-            _priv: (),
-        }
-    }
-}
-
-fn probe_via_mpicc(mpicc: &str) -> std::io::Result<Library> {
-    // Capture the output of `mpicc -show`. This usually gives the actual compiler command line
-    // invoked by the `mpicc` compiler wrapper.
-    Command::new(mpicc).arg("-show").output().map(|cmd| {
-        let output = String::from_utf8(cmd.stdout).expect("mpicc output is not valid UTF-8");
-        // Collect the libraries that an MPI C program should be linked to...
-        let libs = collect_args_with_prefix(output.as_ref(), "-l");
-        // ... and the library search directories...
-        let libdirs = collect_args_with_prefix(output.as_ref(), "-L")
-            .into_iter()
-            .map(PathBuf::from)
-            .collect();
-        // ... and the header search directories.
-        let headerdirs = collect_args_with_prefix(output.as_ref(), "-I")
-            .into_iter()
-            .map(PathBuf::from)
-            .collect();
-
-        Library {
-            libs,
-            lib_paths: libdirs,
-            include_paths: headerdirs,
-            version: String::from("unknown"),
-            _priv: (),
-        }
-    })
-}
-
-/// splits a command line by space and collects all arguments that start with `prefix`
-fn collect_args_with_prefix(cmd: &str, prefix: &str) -> Vec<String> {
-    cmd.split_whitespace()
-        .filter_map(|arg| {
-            if arg.starts_with(prefix) {
-                Some(arg[2..].to_owned())
-            } else {
-                None
-            }
-        })
-        .collect()
-}
-
-/// Probe the environment for an installed MPI library
-pub fn probe() -> Result<Library, Vec<Box<Error>>> {
-    let mut errs = vec![];
-
-    match probe_via_mpicc(&env::var("MPICC").unwrap_or_else(|_| String::from("mpicc"))) {
-        Ok(lib) => return Ok(lib),
-        Err(err) => {
-            let err: Box<Error> = Box::new(err);
-            errs.push(err)
-        }
-    }
-
-    match Config::new().cargo_metadata(false).probe("mpich") {
-        Ok(lib) => return Ok(Library::from(lib)),
-        Err(err) => {
-            let err: Box<Error> = Box::new(err);
-            errs.push(err)
-        }
-    }
-
-    match Config::new().cargo_metadata(false).probe("openmpi") {
-        Ok(lib) => return Ok(Library::from(lib)),
-        Err(err) => {
-            let err: Box<Error> = Box::new(err);
-            errs.push(err)
-        }
-    }
-
-    Err(errs)
 }
