+++ build-probe-mpi/src/os/mod.rs
@@ -1,0 +1,11 @@
+#[cfg(windows)]
+mod windows;
+
+#[cfg(unix)]
+mod unix;
+
+#[cfg(windows)]
+pub use self::windows::probe;
+
+#[cfg(unix)]
+pub use self::unix::probe;
