+++ build.rs
@@ -6,4 +6,9 @@ fn main() {
     if rustc_version::version().unwrap() >= RustcVersion::parse("1.13.0").unwrap() {
         println!("cargo:rustc-cfg=extern_statics_are_unsafe");
     }
+
+    if cfg!(windows) {
+        // Adds a cfg to identify MS-MPI. This should perhaps be a more robust check in the future.
+        println!("cargo:rustc-cfg=msmpi");
+    }
 }
